# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SIAA — Sistema Integrado de Asistencia Académica
  version: 0.1.0
  description: |
    API REST de SIAA. Todos los endpoints requieren HTTPS. Las respuestas de
    error siguen la estructura del catálogo §9.2 del SRS.
  contact:
    name: Equipo SIAA
    email: dev@siaa.edu.co
  license:
    name: Propietario
servers:
  - url: https://api.siaa.edu.co/api/v1
    description: Producción
  - url: http://localhost:8080/api/v1
    description: Desarrollo local

# ─── Seguridad global ──────────────────────────────────────────────────────
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Esquemas de error comunes ─────────────────────────────────────────
  schemas:
    ErrorResponse:
      type: object
      required: [codigo, mensaje, correlationId]
      properties:
        codigo:
          type: string
          example: AUTH_CREDENCIALES_INVALIDAS
          description: Código de error del catálogo §9.2
        mensaje:
          type: string
          example: Credenciales incorrectas
        detalles:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        correlationId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
    FieldError:
      type: object
      required: [campo, error]
      properties:
        campo:
          type: string
          example: correo
        error:
          type: string
          example: required

    # ─── Auth ──────────────────────────────────────────────────────────
...
    EspacioDTO:
      type: object
      required: [id, sedeId, codigo, nombre, capacidad, tipo, estado, nivelValidacion, bufferMetros, versionGeometria, activo, creadoEn, actualizadoEn]
      properties:
        id:
          type: string
          example: 66f0a1b2c3d4e5f60718293c
        sedeId:
          type: string
          example: 66f0a1b2c3d4e5f60718293a
        torre:
          type: string
          example: Torre A
        bloqueId:
          type: string
          example: 66f0a1b2c3d4e5f60718293b
        piso:
          type: integer
          example: 3
        codigo:
          type: string
          example: A-301
        nombre:
          type: string
          example: Aula 301
        capacidad:
          type: integer
          example: 40
        tipo:
          type: string
          enum: [AULA, LABORATORIO, AUDITORIO, TALLER]
          example: AULA
        facultadResponsable:
          type: string
          example: Facultad de Ingeniería
        estado:
          type: string
          enum: [ACTIVO, INACTIVO, MANTENIMIENTO]
          example: ACTIVO
        nivelValidacion:
          type: string
          enum: [AULA, BLOQUE, ZONA]
          example: AULA
        bufferMetros:
          type: number
          example: 10.0
        versionGeometria:
          type: integer
          example: 1
        geometria:
          type: object
          properties:
            type:
              type: string
              example: Polygon
            coordinates:
              type: array
              items:
                type: array
                items:
                  type: array
                  items:
                    type: number
        areaMetrosCuadrados:
          type: number
          example: 54.2
        centroide:
          type: object
          properties:
            type:
              type: string
              example: Point
            coordinates:
              type: array
              items:
                type: number
              example: [-74.08175, 4.60971]
        precisionPromedioMetros:
          type: number
          example: 4.8
        metodoCaptura:
          type: string
          example: RECORRIDO_PERIMETRAL
        activo:
          type: boolean
          example: true
        creadoEn:
          type: string
          format: date-time
        actualizadoEn:
          type: string
          format: date-time
    CrearEspacioRequest:
      type: object
      required: [sedeId, codigo, nombre, tipo]
      properties:
        sedeId:
          type: string
          example: 66f0a1b2c3d4e5f60718293a
        torre:
          type: string
          example: Torre A
        bloqueId:
          type: string
          example: 66f0a1b2c3d4e5f60718293b
        piso:
          type: integer
          example: 3
        codigo:
          type: string
          example: A-301
        nombre:
          type: string
          example: Aula 301
        capacidad:
          type: integer
          example: 40
        tipo:
          type: string
          enum: [AULA, LABORATORIO, AUDITORIO, TALLER]
          example: AULA
        facultadResponsable:
          type: string
          example: Facultad de Ingeniería
        estado:
          type: string
          enum: [ACTIVO, INACTIVO, MANTENIMIENTO]
          example: ACTIVO
        nivelValidacion:
          type: string
          enum: [AULA, BLOQUE, ZONA]
          example: AULA
        bufferMetros:
          type: number
          example: 10.0
    ActualizarEspacioRequest:
      type: object
      properties:
        sedeId:
          type: string
        torre:
          type: string
        bloqueId:
          type: string
        piso:
          type: integer
        codigo:
          type: string
        nombre:
          type: string
        capacidad:
          type: integer
        tipo:
          type: string
          enum: [AULA, LABORATORIO, AUDITORIO, TALLER]
        facultadResponsable:
          type: string
        estado:
          type: string
          enum: [ACTIVO, INACTIVO, MANTENIMIENTO]
        nivelValidacion:
          type: string
          enum: [AULA, BLOQUE, ZONA]
        bufferMetros:
          type: number
        confirmarImpacto:
          type: boolean
          description: Confirmación requerida si el espacio tiene sesiones futuras activas (AC-05).
...
paths:
...
  /espacios:
    get:
      operationId: listarEspacios
      summary: Listar espacios con filtros jerárquicos (AC-01)
      tags: [espacios]
      parameters:
        - name: sedeId
          in: query
          required: false
          schema:
            type: string
        - name: bloqueId
          in: query
          required: false
          schema:
            type: string
        - name: piso
          in: query
          required: false
          schema:
            type: integer
        - name: tipo
          in: query
          required: false
          schema:
            type: string
            enum: [AULA, LABORATORIO, AUDITORIO, TALLER]
        - name: estado
          in: query
          required: false
          schema:
            type: string
            enum: [ACTIVO, INACTIVO, MANTENIMIENTO]
      responses:
        '200':
          description: Lista de espacios obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EspacioDTO'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permiso denegado (requiere aula:leer)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: crearEspacio
      summary: Registrar un nuevo espacio (aula/laboratorio/auditorio/taller)
      tags: [espacios]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearEspacioRequest'
      responses:
        '201':
          description: Espacio creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EspacioDTO'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permiso denegado (requiere aula:crear)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflicto de unicidad de código (AC-03)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Error de validación en los datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /espacios/{id}:
    get:
      operationId: obtenerEspacio
      summary: Obtener detalle de un espacio por ID
      tags: [espacios]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Espacio obtenido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EspacioDTO'
        '404':
          description: Espacio no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      operationId: actualizarEspacio
      summary: Actualizar datos de un espacio (AC-04, AC-05)
      tags: [espacios]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarEspacioRequest'
      responses:
        '200':
          description: Espacio actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EspacioDTO'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permiso denegado (requiere aula:editar)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflicto de código duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Faltan confirmaciones de impacto ante sesiones futuras o datos inválidos (AC-05)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: eliminarEspacio
      summary: Eliminar un espacio con borrado lógico (AC-06)
      tags: [espacios]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Espacio eliminado lógicamente
        '401':
          description: No autenticado
        '403':
          description: Permiso denegado
        '404':
          description: Espacio no encontrado
...
  /bloques/{id}/clonar-piso:
    post:
      operationId: clonarPiso
      summary: Clonar estructura de piso y espacios dentro de un bloque (US-GEO-12)
      tags: [bloques]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pisoOrigen, pisoDestino]
              properties:
                pisoOrigen:
                  type: integer
                pisoDestino:
                  type: integer
      responses:
        '201':
          description: Piso clonado exitosamente
        '400':
          description: Solapamiento o parámetros inválidos
...